inclusão da feature para esconder arquivos

+++ app/src/main/java/br/com/felipe/encrypter/CriptActivity.java
@@ -3,54 +3,53 @@ package br.com.felipe.encrypter;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.os.Bundle;
-import android.os.CountDownTimer;
 import android.support.annotation.NonNull;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.design.widget.TextInputLayout;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.widget.Button;
+import android.widget.EditText;
 import android.widget.Switch;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import com.nbsp.materialfilepicker.MaterialFilePicker;
 import com.nbsp.materialfilepicker.ui.FilePickerActivity;
 
-import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
-import java.io.FileReader;
-import java.io.FileWriter;
-import java.io.IOException;
+import java.io.UnsupportedEncodingException;
+import java.security.InvalidKeyException;
+import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
-import java.util.ArrayList;
-import java.util.List;
 
 import javax.crypto.Cipher;
-import javax.crypto.NoSuchPaddingException;
 import javax.crypto.spec.SecretKeySpec;
+
+import br.com.felipe.utils.Constantes;
 
 public class CriptActivity extends AppCompatActivity {
 
    private Button botao;
    private String modo;
    private Switch botaoSwitch;
-   private static String senha = "0123456789abcdef";
-   private TextInputLayout t;
+   private byte[] senha;
+   private EditText inputSenha;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_cript);
+        //config
         Intent intent = getIntent();
         modo = intent.getStringExtra("modo");
+
+        //init
         botao = findViewById(R.id.selecionar_arquivo);
         botaoSwitch = findViewById(R.id.botao_switch);
-        if(modo.equals("Descriptografar")){
+        inputSenha = findViewById(R.id.textSenha);
+
+
+        if(modo.equals(Constantes.DESCRIPTOGRAFAR_FILE)){
             botaoSwitch.setClickable(false);
             botaoSwitch.setVisibility(View.INVISIBLE);
         }
@@ -59,15 +58,22 @@ public class CriptActivity extends AppCompatActivity {
 
             @Override
             public void onClick(View view) {
+                String password = "0";
+                if(inputSenha.length() > 0){
+                    password = inputSenha.getText().toString();
+                }
+                try {
+                    senha = gerarSenha(password);
+                } catch (Exception e){
+                    e.printStackTrace();
+                }
                 new MaterialFilePicker()
                         .withActivity(CriptActivity.this)
                         .withRequestCode(1000)
-                        .withHiddenFiles(true) // Show hidden files and folders
+                        .withHiddenFiles(true)
                         .start();
             }
         });
-
-
 
     }
     @Override
@@ -78,23 +84,22 @@ public class CriptActivity extends AppCompatActivity {
             String filePath = data.getStringExtra(FilePickerActivity.RESULT_FILE_PATH);
             File file = new File(filePath);
             switch (this.modo){
-                case "Criptografar":{
+                case Constantes.CRIPTOGRAFAR_FILE:{
                     if(botaoSwitch.isChecked()) {
                         filePath = ajustarPath(filePath);
                     }
-                    run(Cipher.ENCRYPT_MODE, file , filePath);
+                    configCripter(Cipher.ENCRYPT_MODE, file , filePath);
                     break;
                 }
 
-                case "Descriptografar":{
-                    run(Cipher.DECRYPT_MODE, file , filePath);
+                case Constantes.DESCRIPTOGRAFAR_FILE:{
+                    configCripter(Cipher.DECRYPT_MODE, file , filePath);
                     break;
                 }
             }
 
         }
-        Intent intent = new Intent(CriptActivity.this, MainActivity.class);
-        startActivity(intent);
+        transicaoActivity();
     }
 
     @Override
@@ -112,39 +117,57 @@ public class CriptActivity extends AppCompatActivity {
     }
 
 
-    private void run(int modoCipher,File origem, String destino){
+    private void configCripter(int modoCipher, File origem, String destino){
+        byte [] result;
         try {
-            //Init
-            FileInputStream inputFile = new FileInputStream(origem);
-            SecretKeySpec key = new SecretKeySpec(senha.getBytes("UTF-8"), "AES");
-            Cipher cipher = Cipher.getInstance("AES");
-            cipher.init(modoCipher, key);
-            byte[] input = new byte[(int) origem.length()];
-
-            //parse
-            inputFile.read(input);
-            inputFile.close();
-            File file;
+            result = execute(modoCipher,origem);
+            if(result == null){
+                transicaoActivity();
+                return;
+            }
             if(!botaoSwitch.isChecked()){
                 origem.delete();
             }
-            file = new File(destino);
-
-            byte[] result = cipher.doFinal(input);
-
-            FileOutputStream outputFile = new FileOutputStream(file);
+            FileOutputStream outputFile = new FileOutputStream(new File(destino));
 
             //save
             outputFile.write(result);
             outputFile.close();
-            Toast toast = Toast.makeText(CriptActivity.this, "Operacação completa!", Toast.LENGTH_SHORT);
-            toast.show();
-        }catch(Exception e){
+            exibirMensagem("Operacação completa!");
+        } catch(Exception e){
             e.printStackTrace();
+            exibirMensagem("Não foi possível realizar a descriptografia");
         }
     }
 
-    public String ajustarPath(String path){
+    private byte[] execute(int modoCipher, File origem) {
+        byte[] input, result = null;
+        try {
+            //configuracoes
+            FileInputStream fileOrigem = new FileInputStream(origem);
+            SecretKeySpec key = new SecretKeySpec(senha, "AES");
+            //set tipo de criptografia
+            Cipher cipher = Cipher.getInstance("AES");
+            //set o modo
+            cipher.init(modoCipher, key);
+
+            input = new byte[(int) origem.length()];
+            fileOrigem.read(input);
+            fileOrigem.close();
+
+            //executa
+            result  = cipher.doFinal(input);
+
+        } catch (InvalidKeyException eW) {
+            exibirMensagem("Senha incorreta!");
+            transicaoActivity();
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return result;
+    }
+
+    private String ajustarPath(String path){
         String paths [] = path.split("/");
         String pathFinal="";
         paths[paths.length-1] ="protegido-"+ paths[paths.length - 1];
@@ -152,6 +175,23 @@ public class CriptActivity extends AppCompatActivity {
             pathFinal = pathFinal + "/"+ s;
         }
         return pathFinal;
+    }
+
+
+    private byte[] gerarSenha(String senha) throws NoSuchAlgorithmException, UnsupportedEncodingException {
+        MessageDigest messageDigest = MessageDigest.getInstance("SHA-256");
+        return messageDigest.digest(senha.getBytes("UTF-8"));
+    }
+
+    private void transicaoActivity(){
+        Intent intent = new Intent(CriptActivity.this, SelectCriptActivity.class);
+        startActivity(intent);
+        finish();
+    }
+
+    private void exibirMensagem(String msg){
+        Toast toast = Toast.makeText(CriptActivity.this, msg, Toast.LENGTH_SHORT);
+        toast.show();
     }
 
 }
